requirements.txt
pandas
openpyxl
psycopg2
python-dotenv